requirements.txt
streamlit==1.28.1
google-generativeai==0.3.2
chromadb==0.4.18
sentence-transformers==2.2.2
python-dotenv==1.0.0
pandas==2.1.3
numpy>=1.26.0
boto3==1.34.14
requests==2.31.0
typing-extensions>=4.8.0
tqdm==4.66.1
langchain==0.1.0
langchain-community==0.0.13
pydantic>=2.5.2
python-magic==0.4.27
tiktoken==0.5.2
tenacity==8.2.3
beautifulsoup4==4.12.2
markdown==3.5.1
PyYAML==6.0.1
watchdog==3.0.0